현재 민원 내용에 대한 내용을 bllossomQ8 모델에서 [민원요지]받아오고, 해당 [민원 요지랑] 템플릿, 및 [답변 요지]를 조합. -> 추후 실 사용시 민원 요지만 받아서 사용할 계획
사용자가 입력한 [답변 요지]를 벡터 db(answer_yogi) 에서 찾아 유사하면. 해당 답변요지의, 답변 결과 를참고하여  ,  답변생성. 일정 유사도 벗어나면 . 동일한 유사 답변 없음 출력.

더 많은 데이터로 테스트 필요.  여러가지 민원 및 이전 답변 생성 임의로 만들어
테스트더 할 예정 .

현재 까지 하나의 유사한 민원 데이터 만 참고 가능 .

추후 db 속성에 answer_yogi 추가 해야 할수도 잇습니다.

파일 위치 수정
